KAFKA-12892: Use dedicated root in ZK ACL test (#10821)
Having the `testChrootExistsAndRootIsLocked` test in a separate `ZookeeperTestHarness` isn't enough to prevent the ACL changes to the ZK root from affecting other integration tests. So instead, let's use a dedicated znode for this test. It still works because `makeSurePersistentPathExists` uses `createRecursive`, which will recurse and act the same for the root or a given znode.

 Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>
